+++ fluentgen/src/main/java/com/azure/autorest/fluent/FluentGen.java
@@ -14,7 +14,7 @@ import com.azure.autorest.fluent.mapper.FluentMapperFactory;
 import com.azure.autorest.fluent.template.FluentTemplateFactory;
 import com.azure.autorest.fluent.transformer.FluentTransformer;
 import com.azure.autorest.fluent.util.FluentJavaSettings;
-import com.azure.autorest.fluent.util.FluentNamerFactory;
+import com.azure.autorest.fluent.namer.FluentNamerFactory;
 import com.azure.autorest.mapper.Mappers;
 import com.azure.autorest.model.clientmodel.Client;
 import com.azure.autorest.model.clientmodel.ClientException;
